# Conflicts:
#	doc/src/Commands_pair.rst
#	src/ELECTRODE/electrode_vector.cpp
#	src/ELECTRODE/fix_electrode_conp.cpp
#	unittest/force-styles/CMakeLists.txt
#	unittest/force-styles/test_config.h
#	unittest/force-styles/test_config_reader.cpp
#	unittest/force-styles/test_config_reader.h
#	unittest/force-styles/test_main.cpp
#	unittest/force-styles/test_main.h

+++ src/EXTRA-FIX/fix_ttm_thermal.cpp
@@ -293,7 +293,7 @@ void FixTTMThermal::end_of_step()
   double inner_dt = update->dt;
   double voxel_coeff =(1.0/dx/dx + 1.0/dy/dy + 1.0/dz/dz);
 
-  std::vector<double> grid_fourier(nzgrid * nygrid * nxgrid);
+  std::vector<double> grid_fourier((size_t) nzgrid * nygrid * nxgrid);
   int index = 0;  // Location unimportant, only max value
     for (iz = 0; iz < nzgrid; iz++)
       for (iy = 0; iy < nygrid; iy++)
